asset_buyer/requirements.txt
Flask==1.0.3
python-bitcoinrpc==1.0